@@ -11,7 +11,7 @@
   },
   "dependencies": {
     "core-js": "^3.6.5",
-    "vue": "^2.6.11"
+    "vue": "^3.0.11 "
   },
   "devDependencies": {
     "@babel/core": "^7.14.3",
@@ -26,6 +26,8 @@
     "babel-loader": "^8.2.2",
     "eslint": "^6.7.2",
     "eslint-plugin-vue": "^6.2.2",
+    "node-sass": "^6.0.0",
+    "sass-loader": "^10",
     "vue-loader": "^15.9.7",
     "vue-template-compiler": "^2.6.11"
   },
